requirements.txt
flask==3.0.0
google-generativeai==0.8.3
gunicorn==21.2.0
werkzeug==3.0.1